Adding dev jwt token to replace vault access
